스킬테이블테스트완료, 스킬선택UI 추가

+++ DungeonCleaner/Assets/Scripts/Skills/SkillData.cs
@@ -10,5 +10,5 @@ public class SkillData : ScriptableObject
     public float duration;
     public float tickInterval;
     public int projectileCount;
-    public int projectileSpeed;
+    public float projectileSpeed;
 }
